pyproject.toml
[tool.poetry]
name = "repayment"
version = "0.1.0"
description = ""
authors = ["pp-ghw <huawaai@peerpower.co.th>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
